refactor: #12 MyPageController 이름 변경 및 API 정리
